requirements.txt
fastapi
requests
uvicorn
python-dotenv
pydantic
langchain
langchain-community
langchain-groq